+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 0.5.41 (Pending):
+
 ## 0.5.40:
 * Send a DataLoadCompleteEvent when a data load completes.
 * Send a SystemErrorEvent on the EventBus when a system error occurs.